Remove separate Unicode width table header
